Many changes made.
Added a custom shape to the list.  The custom shape is made from an array saved with pickle.  The array contains x,y and z
coordinates and is in the hardcoded file named GCode-array.pik .

Added a new python program that helps drawing the custom shape ... which for a meringue should be simple,  so the
program (meringue_draw_gcode_array.py) is rather basic. It saves to the GCode-array.pik file.

Made a rudimentary meringue icon that can be associated with the meringue_gcode_GUI.py th start the program
from a desktop.